Feat: Beverages ImageView, addToStock buttons
