+++ VendingMachineApp/VendingMachineApp/Controllers/ViewController.swift
@@ -12,6 +12,12 @@ class ViewController: UIViewController {
     var vendingMachine = VendingMachine(products: [])
 
     @IBOutlet var productsStackView: [UIStackView]!
+    @IBOutlet var addToStockButtons: [UIButton]!
+    @IBOutlet var beverageImageViews: [UIImageView]!
+    
+    @IBAction func addToStockButtonTouched(_ sender: Any) {
+        let button = sender as! UIButton
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -20,6 +26,9 @@ class ViewController: UIViewController {
     }
     
     private func setupUI() {
+        beverageImageViews.forEach {
+            $0.layer.cornerRadius = 16
+        }
     }
 }
