+++ VendingMachineApp/VendingMachineApp/Views/OldProductStackView.swift
@@ -1,0 +1,57 @@
+//
+//  ProductStackView.swift
+//  VendingMachineApp
+//
+//  Created by Cory Kim on 2020/02/25.
+//  Copyright © 2020 corykim0829. All rights reserved.
+//
+
+import UIKit
+
+class OldProductStackView: UIStackView {
+    
+    let addToStockButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.setTitle("추가", for: .normal)
+        button.backgroundColor = .green
+        button.setTitleColor(.white, for: .normal)
+        return button
+    }()
+
+    let productImageView: UIImageView = {
+        let imageView = UIImageView(image: #imageLiteral(resourceName: "bananaMilk"))
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        imageView.heightAnchor.constraint(equalToConstant: 120).isActive = true
+        imageView.widthAnchor.constraint(equalTo: imageView.heightAnchor).isActive = true
+        imageView.contentMode = .scaleAspectFit
+        imageView.backgroundColor = .white
+        imageView.layer.cornerRadius = 16
+        return imageView
+    }()
+
+    let stockCountLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = .white
+        label.text = "0"
+        label.textAlignment = .center
+        label.backgroundColor = .red
+        return label
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configure()
+    }
+    
+    required init(coder: NSCoder) {
+        super.init(coder: coder)
+        configure()
+    }
+    
+    private func configure() {
+        translatesAutoresizingMaskIntoConstraints = false
+        axis = .vertical
+        spacing = 8
+        [addToStockButton, productImageView, stockCountLabel].forEach{ addArrangedSubview($0) }
+    }
+}
