+++ VendingMachineApp/VendingMachineApp/Views/ProductsStackView.swift
@@ -1,0 +1,21 @@
+//
+//  ProductsStackView.swift
+//  VendingMachineApp
+//
+//  Created by Cory Kim on 2020/02/28.
+//  Copyright © 2020 corykim0829. All rights reserved.
+//
+
+import UIKit
+
+class ProductsStackView: UIStackView {
+
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
